fix bug

+++ dsHoaDon.h
@@ -54,7 +54,7 @@ hd_Node* MakeHdNode(hoa_don data);
 void AddLastListHD(ds_hoa_don*& dshd, hoa_don hd);
 void InputFormHD(string ct[], int sl, int length);
 hd_Node* FindSoHDExist(ds_hoa_don* dshd, string ID);
-int ComputeValue(int sl, double donGia, double VAT);
+double ComputeValue(int sl, double donGia, double VAT);
 int PickHDType();
 void ButtonFunction();
 void TableOfHD();
